+++ src/Exception/AlreadyInstalledException.php
@@ -3,7 +3,7 @@
 /*
  * This file is part of the Asplode package.
  *
- * Copyright © 2013 Erin Millard
+ * Copyright © 2014 Erin Millard
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
@@ -11,23 +11,24 @@
 
 namespace Eloquent\Asplode\Exception;
 
-use Exception as NativeException;
+use Exception;
 
 /**
  * This Asplode instance has already been installed.
  */
-final class AlreadyInstalledException extends LogicException
+final class AlreadyInstalledException extends Exception implements
+    AsplodeExceptionInterface
 {
     /**
      * Construct a new already installed exception.
      *
-     * @param NativeException|null $previous The previous exception, if
-     *     available.
+     * @param Exception|null $previous The cause, if available.
      */
-    public function __construct(NativeException $previous = null)
+    public function __construct(Exception $previous = null)
     {
         parent::__construct(
             'This instance of Asplode has already been installed.',
+            0,
             $previous
         );
     }