+++ src/Exception/AsplodeExceptionInterface.php
@@ -3,7 +3,7 @@
 /*
  * This file is part of the Asplode package.
  *
- * Copyright © 2013 Erin Millard
+ * Copyright © 2014 Erin Millard
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
@@ -12,8 +12,9 @@
 namespace Eloquent\Asplode\Exception;
 
 /**
- * Interface implemented by all Asplode-related exceptions.
+ * Interface use to identify all Asplode-related exceptions, excluding actual
+ * error exceptions.
  */
-interface Exception
+interface AsplodeExceptionInterface
 {
 }