+++ test/suite/Exception/ErrorHandlingConfigurationExceptionTest.php
@@ -3,7 +3,7 @@
 /*
  * This file is part of the Asplode package.
  *
- * Copyright © 2013 Erin Millard
+ * Copyright © 2014 Erin Millard
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
@@ -11,22 +11,17 @@
 
 namespace Eloquent\Asplode\Exception;
 
-use Exception as NativeException;
-use Phake;
+use Exception;
 use PHPUnit_Framework_TestCase;
 
-class LogicExceptionTest extends PHPUnit_Framework_TestCase
+class ErrorHandlingConfigurationExceptionTest extends PHPUnit_Framework_TestCase
 {
     public function testException()
     {
-        $previous = new NativeException;
-        $exception = Phake::partialMock(
-            __NAMESPACE__.'\LogicException',
-            'foo',
-            $previous
-        );
+        $previous = new Exception;
+        $exception = new ErrorHandlingConfigurationException($previous);
 
-        $this->assertSame('foo', $exception->getMessage());
+        $this->assertSame('Error handling is incorrectly configured.', $exception->getMessage());
         $this->assertSame(0, $exception->getCode());
         $this->assertSame($previous, $exception->getPrevious());
     }